refactor(handlers): split the registry god-module into domain modules
registry.py was 2,237 lines and growing with every rule. Physical split
only - the dispatch contract, public names, and behavior are identical:

- registry.py (135 lines) keeps registration/dispatch (HandlerRegistry,
  generic_handler) and re-exports every public name for back-compat.
- Domain modules compose as mixins (handlers never used self):
  generic, dependencies, runtime, monitoring, deployment, bindings,
  project, durable, integrations; module-level evaluators moved with
  their domains.
- Test patch targets repointed from registry to the owning domain
  module (same objects, new lookup namespace).
- docs/handlers.md and docs/architecture.md describe the new layout.

Verified: full suite green, coverage 96.97% (>=95%), mypy strict 65
files clean, ruff clean, pin linters pass, and a live doctor run on the
http-trigger example produces identical section/item/status counts.

Closes #387
