+++ tests/test_handler.py
@@ -164,7 +164,7 @@ def test_compare_func_core_tools_version_pass(monkeypatch: MonkeyPatch) -> None:
     from azure_functions_doctor import handlers
 
     monkeypatch.setattr(
-        handlers.registry,
+        handlers.generic,
         "resolve_target_value",
         lambda t: "4.0.5455" if t == "func_core_tools" else "",
     )
@@ -186,7 +186,7 @@ def test_compare_func_core_tools_version_fail_not_installed(monkeypatch: MonkeyP
     from azure_functions_doctor import handlers
 
     monkeypatch.setattr(
-        handlers.registry,
+        handlers.generic,
         "resolve_target_value",
         lambda t: "not_installed" if t == "func_core_tools" else "",
     )
@@ -208,7 +208,7 @@ def test_compare_func_core_tools_version_fail_old_version(monkeypatch: MonkeyPat
     from azure_functions_doctor import handlers
 
     monkeypatch.setattr(
-        handlers.registry,
+        handlers.generic,
         "resolve_target_value",
         lambda t: "3.0.0" if t == "func_core_tools" else "",
     )
